+++ src/main/java/uni/climatemonitor/graphics/DetailsPage.form
@@ -47,7 +47,7 @@
                     </constraints>
                     <properties>
                       <autoscrolls value="false"/>
-                      <font size="20" style="1"/>
+                      <font size="24" style="1"/>
                       <horizontalAlignment value="0"/>
                       <text value=""/>
                     </properties>
@@ -74,7 +74,7 @@
                 </children>
               </grid>
               <grid id="6a4bd" binding="ParametersContainer" layout-manager="GridLayoutManager" row-count="8" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-                <margin top="20" left="20" bottom="0" right="0"/>
+                <margin top="30" left="20" bottom="0" right="0"/>
                 <constraints>
                   <grid row="1" column="0" row-span="1" col-span="2" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
                 </constraints>
@@ -558,13 +558,6 @@
                   <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
                 </constraints>
               </vspacer>
-              <vspacer id="a122c">
-                <constraints>
-                  <grid row="1" column="1" row-span="1" col-span="2" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false">
-                    <preferred-size width="-1" height="200"/>
-                  </grid>
-                </constraints>
-              </vspacer>
               <grid id="21615" binding="ButtonsPnl" layout-manager="FlowLayout" hgap="5" vgap="0" flow-align="2">
                 <constraints>
                   <grid row="0" column="0" row-span="1" col-span="3" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
@@ -627,6 +620,11 @@
                   </grid>
                 </children>
               </grid>
+              <vspacer id="a4405">
+                <constraints>
+                  <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
+                </constraints>
+              </vspacer>
             </children>
           </grid>
         </children>
